Projekt 1/requirements.txt
ipykernel
jupyter
numpy==1.26.4
matplotlib>=3.8.0
Pillow>=10.0.0
streamlit==1.54.0